refactor: extract ttv scoring helpers
