+++ cmat/ttv_sim.py
@@ -5,6 +5,8 @@ from matplotlib import pyplot as plt
 from multiprocessing import get_context
 from tqdm.auto import tqdm
 
+from .scoring import get_chi2, get_chi2_v, get_rms, get_rms_v
+
 mj_to_ms = 9.5e-4
 me_to_ms = 3.0e-6
 rj_to_rs = 0.102792236
@@ -23,31 +25,6 @@ systems by evaluating the Mean Exponential Growth of Nearby Orbits (MEGNO)
 values. This indicator can help determine the chaotic behavior of a system
 and find potentially stable configurations.
 """
-
-
-def get_chi2(ttv_rebound, epoch, ttv_mcmc, ttv_err):
-    # print(ttv_rebound)
-    rangea = range(epoch[-1] - epoch[0])
-    T = [ttv_rebound[np.array(epoch - epoch[0]) + a] for a in rangea]
-    chi2 = (((T - ttv_mcmc) ** 2) / ttv_err**2).sum(axis=1)
-    return chi2.min()
-# This function calculates the chi-squared value between the TTVs generated
-# by the rebound integrator and the TTVs obtained from MCMC fitting.
-# The chi-squared value is a measure of the goodness of fit.
-
-
-def get_rms(ttv_rebound):
-    rms = np.sqrt(np.mean(ttv_rebound**2))
-    return rms
-# This function calculates the root-mean-square (RMS) value of the TTVs
-# generated by the rebound integrator.
-
-
-get_chi2_v = np.vectorize(
-    get_chi2, excluded=["epoch", "ttv_mcmc", "ttv_err"], signature="(n)->()"
-)
-get_rms_v = np.vectorize(get_rms, signature="(n)->()")
-# Vectorize the get_chi2 and get_rms functions for faster computation
 
 
 class ttv_sim:
